Log critical memory usage

+++ 3dEngine/src/scene/Scene.h
@@ -77,7 +77,7 @@ namespace urchin {
             int fpsLimit;
 
             //renderer
-            std::uint64_t frameIndex;
+            std::uint32_t frameIndex;
             ScreenRender screenRenderTarget;
             std::vector<std::unique_ptr<Renderer3d>> renderers3d;
             Renderer3d* activeRenderer3d;
