+++ 3dEngine/src/scene/renderer3d/lighting/shadow/ShadowManager.cpp
@@ -245,7 +245,7 @@ namespace urchin {
         #endif
     }
 
-    void ShadowManager::updateShadowMaps(std::uint64_t frameIndex, unsigned int numDependenciesToShadowMaps) const {
+    void ShadowManager::updateShadowMaps(std::uint32_t frameIndex, unsigned int numDependenciesToShadowMaps) const {
         ScopeProfiler sp(Profiler::graphic(), "updateShadowMap");
         unsigned int renderingOrder = 0;
