+++ 3dEngine/src/scene/renderer3d/postprocess/antialiasing/AntiAliasingApplier.cpp
@@ -113,7 +113,7 @@ namespace urchin {
         return config;
     }
 
-    void AntiAliasingApplier::applyAntiAliasing(std::uint64_t frameIndex, unsigned int numDependenciesToAATexture) const {
+    void AntiAliasingApplier::applyAntiAliasing(std::uint32_t frameIndex, unsigned int numDependenciesToAATexture) const {
         ScopeProfiler sp(Profiler::graphic(), "applyAA");
 
         renderTarget->render(frameIndex, numDependenciesToAATexture);
