+++ 3dEngine/src/scene/renderer3d/postprocess/antialiasing/AntiAliasingApplier.h
@@ -29,7 +29,7 @@ namespace urchin {
             void updateConfig(const Config&);
             const Config& getConfig() const;
 
-            void applyAntiAliasing(std::uint64_t, unsigned int) const;
+            void applyAntiAliasing(std::uint32_t, unsigned int) const;
 
         private:
             struct AntiAliasingShaderConst {
