+++ 3dEngine/src/texture/filter/TextureFilter.h
@@ -22,7 +22,7 @@ namespace urchin {
 
             const std::shared_ptr<Texture>& getTexture() const;
 
-            void applyFilter(std::uint64_t, unsigned int) const;
+            void applyFilter(std::uint32_t, unsigned int) const;
 
         protected:
             virtual std::string getShaderName() const = 0;
